Add deployment options to hugo.yaml
Note that I couldn't get the r2 deployment to work. I _think_ this is because r2 is not returning correct hashes for the files in the bucket, and thus the goblob code can't figure out which files to replace.  This might be an issue I can't fix, but it requires more research.