Remove disableSSL flag from r2 URL
This was causing issues with hugo deploy where Cloudflare was returning 301 from a PUT request, and the aws sdk (probably) used by the gocloud.dev/blob library would then attempt to GET instead of PUT at the new address.  This caused a mismatch of x-amz-content-sha256 used to validate requests, and the upload failed (never tried to PUT at the new address)

Removing this parameter allows it to use SSL from the beginning, and thus not get caught in the redirect error.  It could be a bug in the AWS SDK (should not try to GET after getting a 301 on a PUT) or from Cloudflare R2 (should return a 308 instead of 301 which is apparently stricter about not changing the HTTP verb)

I don't know and I don't care

https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/301
https://gocloud.dev/howto/blob/#s3-compatible
https://docs.aws.amazon.com/AmazonS3/latest/API/sig-v4-header-based-auth.html